AC-985 : Update operation Fixed up
1. removed the clutter and fixed the update operation using sync operation.
